dom/base/test/test_input_vsync_alignment_lower_than_normal.html

<!DOCTYPE HTML>
<html>
<head>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body >
  <script type="text/javascript">
  SimpleTest.waitForExplicitFinish();

  let runOrder = [];
  function checkResult() {
    if (runOrder.length == 3) {
      if (runOrder[0] === "Normal") {
        isDeeply(
            runOrder,
            ["Normal", "InputHigh", "Vsync"],
            "Input priority tasks should let normal tasks to run first when there's no pending vsync"
        );
        SimpleTest.finish();
      } else {
        runOrder = [];
        runTest();
      }
    }
  }

  function runTest() {
    window.requestAnimationFrame(() => {
      window.requestAnimationFrame(() => {
        runOrder.push("Vsync");
        checkResult();
      });
      SpecialPowers.Services.tm.dispatchToMainThread(function() {
        runOrder.push("InputHigh");
        checkResult();
      }, SpecialPowers.Ci.nsIRunnablePriority.PRIORITY_INPUT_HIGH);
      SpecialPowers.Services.tm.dispatchToMainThread(function() {
        runOrder.push("Normal");
        checkResult();
      }, SpecialPowers.Ci.nsIRunnablePriority.PRIORITY_NORMAL);
    });
  }

  SpecialPowers.pushPrefEnv({
      set: [["dom.input_events.strict_input_vsync_alignment", true]]
  }, runTest());
  </script>
</body>
</html>
